% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ardecheinv}
\alias{ardecheinv}
\title{Temporal data}
\format{The \code{ardecheinv} is a list of 3 components.

 \code{env}
   is a dataframe of 67 rows with 5 environmental tables, collected in spring and autumn.


 \code{Invertebrates}
   is a dataframe of 67 rows and 57 species of invertebrates.


 \code{code}
   is a dataframe with 57 rows and 2 columns, the species scientific name and their respective code.}
\source{
Merigoux, S. and Doledec, S. (2004). Hydraulic requirements of stream communities: A case study on invertebrates. \emph{Freshwater Biology}, \bold{49}(5), 600-613.
}
\usage{
data(ardecheinv)
}
\description{
The \code{ardecheinv} data are the temporal example used as an ecological application of the WitOMI calculations in
Karasiewicz et al. (2017).
}
\references{
Karasiewicz S.,Doledec S.and Lefebvre S. (2017). Within outlying mean indexes: refining the OMI analysis for the realized niche decomposition. \emph{PeerJ} 5:e3364. \url{https://doi.org/10.7717/peerj.3364}.
}
\keyword{datasets}
