% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drome}
\alias{drome}
\title{Spatial data}
\format{
The \code{drome} is a list of 3 components.

 \code{env}
   is a dataframe with 64 rows with 6 environmental tables, collected in 10 different rivers.


 \code{fish}
   is a dataframe with 64 rows and 13 columns (12 fish species, including young and older trouts).


 \code{code}
   is a dataframe with 13 rows and 2 columns, the species, common and scientific, name and their respective code.
}
\source{
Doledec S., Chessel D. and Gimaret C. (2000). Niche separation in community analysis: a new method. \emph{Ecology},\bold{81}(10), 2914-1927.
}
\usage{
data(drome)
}
\description{
The \code{drome} data are the spatial example used as an ecological application of the WitOMI calculations in
Karasiewicz et al. (2017).
}
\references{
Karasiewicz S.,Doledec S.and Lefebvre S. (2017). Within outlying mean indexes: refining the OMI analysis for the realized niche decomposition. \emph{PeerJ} 5:e3364. \url{https://doi.org/10.7717/peerj.3364}.
}
\keyword{datasets}
