% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margvect.R
\name{margvect}
\alias{margvect}
\title{Summarizes the results of the sub-environmental spaces}
\usage{
margvect(x, xax = 1, yax = 2, colo = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{subniche}.}

\item{xax}{column for abscisse.}

\item{yax}{column for ordinate.}

\item{colo}{string of colors of equal length than the number of G_k, if NULL "red"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The function plot the resulting sub-environmental space of the WitOMI calculation.
}
