% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepfactorrow.R
\name{sep.factor.row}
\alias{sep.factor.row}
\title{A function to seperate a matrix, by row, into submatrices.}
\usage{
sep.factor.row (x,factor)
}
\arguments{
\item{x}{a matrix.}

\item{factor}{a factor of the same length as the number of row in the matrix.}
}
\value{
list of submatrices
}
\description{
separate matrix by rows into submatrices
}
